Functorise Ast_to_nast by creating Ast_to_aast
Summary:
For the unified AST project, we need to quickly go from a legacy AST to a concrete version of the AAST (either NAST or TAST). Originally, we went from an AST to a NAST, however for compilation, we need a TAST. This resulted in the following flow for compilation:

AST -> NAST -> TAST

Going via an NAST is unnecessary, so this diff functorises Ast_to_nast by creating a new module, Ast_to_aast. This means we can go directly to a TAST from an AST.

Differential Revision: D15028936

fbshipit-source-id: 16a71484634677e0004b9ee59717bc8647adad07