Include classes not known to exist in the linearization
Summary:
Currently, it is always an error to extend a class not known to exist. This wasn't always the case, though--until {D14037113}, when the assume_php setting was enabled, the typechecker would allow unbound names in partial-mode files in order to aid in gradual typing. In addition to not emitting errors for unbound names specified as an ancestor, we would include these names in ancestor collections for uses like subtyping.

For example, if we enabled assume_php, and wrote these definitions in a partial-mode file:

```
<?hh // partial

// If the assume_php setting is enabled,
// and no class named DefinedInPhp is
// visible to the typechecker, it is not
// an error to extend it in partial mode.
class MyClass extends DefinedInPhp {}

// Similarly, we can use the unbound
// class name in a type annotation without
// an error:
function f(DefinedInPhp $x): void {}
```

The typechecker does some (limited) enforcement for them in a strict-mode file:

```
<?hh // strict

function test(MyClass $x): void {
  // The typechecker accepts this because
  // "DefinedInPhp" is known to be an
  // ancestor of MyClass, even though it
  // is not known to exist. Classes which
  // do not have DefinedInPhp in their
  // collection of ancestors are not
  // subtypes of DefinedInPhp.
  f($x);

  // Because the typechecker does not
  // know what members DefinedInPhp
  // has, it permits this call.
  $x->foo();
}
```

The use of `$x` in `f` was permitted because the unbound class name would be included in the `dc_ancestors` collection used in subtype tests. Calling `$x->foo()` was permitted because we would track whether any such class was present in a class hierarchy with `dc_members_fully_known`.

We removed the assume_php setting in D14037113, but remnants of support for this setting still exist in the typechecker (in particular, the `members_fully_known` API and its use sites).

To proceed with implementing shallow_class_decl, we need to either build implementations for APIs like `members_fully_known` or remove remaining implementation details related to the `assume_php` feature. It seems to me that the latter is the easier path forward for now--it only requires that we include in the linearization any classes not known to exist. This diff makes that change.

Reviewed By: vassilmladenov

Differential Revision: D15042979

fbshipit-source-id: 64d0d12baaba30b3c0d3a98a90346500a003f2a9