Compute members_not_found lazily for shallow_class_decl
Summary:
This bool is true when no class in a hierarchy declared a class not known to the typechecker to be an ancestor (i.e., there is no ancestor invisible to the typechecker which may have defined members the typechecker cannot know about).

After {D15042979}, this is very easy to express using the linearization--we can simply say "for all elements in the linearization, none has the class_not_found flag".

Reviewed By: vassilmladenov

Differential Revision: D15043165

fbshipit-source-id: 64c8d5cd18f3fd0359893a9da68cd1a951c2e3a1