Look up parent and trait ancestors lazily from the linearization
Summary:
In legacy decl, the `Typing_classes_heap.extends` API provides access to the `tc_extends` collection, which, after {D14986873}, is identical to the `dc_extends` collection, which contains the names of all extends-ancestors and traits used in a class's hierarchy. In the typechecker, it is used only to control access to protected members (which is [why we include traits in this collection](diffusion/FBS/browse/master/fbcode/hphp/hack/src/decl/decl.ml;39b9579563d6f9c6385a7bf294d09e42c0e73434$132-134)). It is otherwise used only in linters, to find classes which extend some class of interest (the fact that traits are included in this collection doesn't seem to be relevant in any of these linters).

Because we already linearize all extends-ancestors, traits, and implemented interfaces with the Ancestor_types linearization kind, and the interfaces are marked with mro_consts_only, we can easily reuse this linearization to reproduce the dc_extends collection lazily.

Reviewed By: vassilmladenov

Differential Revision: D15042980

fbshipit-source-id: 5e3b1c4b5d0583447c651b16fbe0eeb5c42d8c53