order the awaits pulled out by position in file.
Summary: Today await lifting depends on the order of the full_fidelity_ast.ml traversal, which in some cases isn't a DFS. Move to sorting the awaits by Position to ensure that there is a well defined order for awaits being run. In the event of Pos.none we assert.

Reviewed By: paulbiss

Differential Revision: D15028472

fbshipit-source-id: 8923872801b8e8891ff47616d261f62066fbb530