Split index builder into library vs executable
Summary:
The goal of this diff stack is to enable hh_global_index_builder to be packaged within the regular HH release, so that it is regularly shipped.

To do that, we're going to split the program into a library (that does the work) and an executable (that parses the command line options).

Ideally, if we choose to, we can do away with the separate executable over time and just simply call the index builder as a method whenever necessary.

Reviewed By: pittsw

Differential Revision: D14999005

fbshipit-source-id: 53c10a9fab957f5679327874b39a5d816eb4796b