Invalid memoized param error should point to the param as the primary position
Summary: At the moment, the `invalid_memoized_param` error has as the primary position the definition of the type of the invalid parameter, which is confusing because that type and the definition of the memoized function can be in different files.  The error should really point to the invalid parameter position.  (This was flagged by D15079343.)

Reviewed By: andrewjkennedy

Differential Revision: D15079438

fbshipit-source-id: 6c5743b1240abec35c4762509e248d8fbfaec22c