Generalize NodeType framework not to depend on Syntax (too limiting)
Summary:
In the Rust implementation, all parser instances need to implement NodeType,
which is not possible if there's one-to-*many* mapping between
constructed nodes (`S::R`) and parsed type (`SyntaxKind`).  For example,
in FactsSmartConstructors, S=FlattenSmartConstructors and R=Node,
consider mapping between Node::Ignored and SyntaxKind::Missing.

Therefore, port the SmartConstructorWrappers.SyntaxKind wrapper to Rust,
essentially converting type S::R to (SyntaxKind, S::R), so that it
can *automatically* track syntax kind as it builds the result type R.
This removes the burden on the users and enables FactsSmartConstructors
implementation in Rust to match the one in OCaml.

Reviewed By: dabek

Differential Revision: D15023646

fbshipit-source-id: fce61761365008f0b082ab4d991384ccf0e306f5