Add FPushClsMethodR* as an optimizations for ReifiedName + FPushClsMethod*
Summary: This diff adds a new bytecode to optimize the combination of `ReifiedName` + `FPushClsMethod` since for reified generics we do know the name of the function statically but using `FPushClsMethod` was preventing us from using this static knowledge. Also, according to strobelight data, most of the time is spend on generating the string from reified generics but in this case we do not need to do that, so this diff elides this string creation.

Reviewed By: jano

Differential Revision: D14932066

fbshipit-source-id: b31c7b4bae6a07e1fedb6c3c4cb4fde42e6bbe69