transform subtype assertions involving Tany into subtype assertions on inner types ("propagate Tany")
Summary:
In Typing_subtype, for a subtype assertion involving Tany on one side and a type with inner types on the other side, e.g. generics, functions, transform that assertion into assertions on the inner types.

To do that, match each inner type to Tany, for example map `C<D>` to `C<_>` then transform
```
C<D> <: _
```
into
```
C<D> <: C<_>
```
which should result on a subtype assertion involving `D` and `_` (depending on the variance of `C` here).

Reviewed By: andrewjkennedy

Differential Revision: D14973783

fbshipit-source-id: 8d643c75308d3dc54482aba2fb138d154c8a66a1