Fix type-checking of calls to parent::__construct in the presence of where constraints
Summary:
Consider the following example (added as a new test case below):
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

interface I {
  abstract const type T2;
}

abstract class A<T1 as I, T2> {
  public function __construct(T1 $_) where T2 = T1::T2 {}
}

final class B implements I {
  const type T2 = int;
}

final class C extends A<B, B::T2> {
  public function __construct(B $b) {
    parent::__construct($b);
  }
}
```
Both legacy and new type inference report an error here:
```
You are trying to access the member T2 but this is not an object, it is a value of generic type T1 (Typing[4062])
```
(the error produced by the new inference has more extra lines).  The problem is that when type-checking calls to the method `parent::__construct`, the latter is localized using an empty substitution for generic type parameters.

Reviewed By: andrewjkennedy

Differential Revision: D15097797

fbshipit-source-id: 361ecf5240d72f34e9fe33a6c9b758d555a40245