Make SyntaxSmartConstructors "inheritable" & advance state
Summary:
In OCaml, `SyntaxSmartConstructors` is a reusable piece
of code that abstracts over a `Syntax` and delegates `make_`
calls to Syntax's make_ methods, with the difference that Syntax SC
also has the State parameter through which the state (typed `State::T`)
is advanced via `State::next`.  Implementors then override a subset
of these "included" methods.  In Rust, make Syntax SC a trait
(instead of a concrete type) with default methods; this way,
implementors can override specific methods by redefining them, e.g.:

    impl SyntaxSmartConstructors<MySyntax, MyState>
      for MySmartConstuctors {   // param. MyState is optional
      ...
      fn make_FOO(...) -> (MyState::T, Self::R) { ... }
    }

The delegation from SmartConstructors to SyntaxSmartConstructors is
currently generated but is easily replaced with macros in the future.

Reviewed By: Wilfred

Differential Revision: D14837542

fbshipit-source-id: d9ad6ff44e2b535f5de74dd0dfa6bdae7e5c3563