Fix effect_free condition for SetL, PopL and UnsetL
Summary:
There are three main states when we're interping.

 - analyzing the function
 - analyzing an inline invocation of the function
 - speculating a block to determine if we can skip over it

In the first two, modifications of locals should be "effect free",
because they have no visible effects from the point of view of
callers.

So SetL, PopL and UnsetL (PushL is implemented in terms of CGetL and
UnsetL) should be effect_free in the first two, but not in the third.

We were only marking them effect_free in the second case, which could
cause us to fail to fold some effect_free functions.

Reviewed By: ricklavoie

Differential Revision: D15103975

fbshipit-source-id: 68655a4b8b417c14571db51d2d929fbcdbbc6d5e