Allow pops inserted by speculate to be optimized
Summary:
If speculate determines that it can skip over some blocks, but also
needs to pop some elements off the stack to match what those blocks
would have done, it just inserts the pops at the end of the block.

If the pops were preceded by poppables, or other instructions that
could interact with the pops, we don't optimize. Often thats not a
problem, because we end up revisiting the block, but in some cases we
end up with sequences like 'True; PopC'.

Preserve the state at the end of the original block so we can just
interpStep the pops, and let that take care of optimizing the block.

Reviewed By: ricklavoie

Differential Revision: D15111441

fbshipit-source-id: 3a35e6deb4427696a9db3477f3288c1350e08422