Simplify ObjectData allocation in APC
Summary:
In apc-object.cpp, when m_persistent is true, then the value of m_fast_init is always equal to !cls->instanceCtor(). m_fast_init is only read when m_persistent is true and we've entered APCObject::createObject; we have the class available there, so we don't need to store m_fast_init.
We use m_fast_init in part so that we can punt from calling ObjectData::newInstanceNoPropInit to calling ObjectData::newInstance, because the former doesn't support allocating objects with an instanceCtor and asserts as much.
This is the only call site of newInstanceNoPropInit, so just make it work for classes with an instanceCtor and remove the need to punt.
Rather than copy-pasta more out of newInstance into newInstanceNoPropInit, move their common bits into newInstanceImpl.

Reviewed By: markw65

Differential Revision: D15045070

fbshipit-source-id: 78a9f39a7d8263afcd10445c5871674f384a3f83