Simplify ObjectData allocation in CreateCl
Summary:
In the interpreter, there's no need for CreateCl to go through newInstance: closures never have reified generics, closure classes never needInitialization(), and we know we're always going to go into the instanceCtor path. Just call the instanceCtor directly instead.
Similarly, in the jit, we never needInitialization() and we're always going to end up in the ConstructInstance path, so just gen that directly.

Reviewed By: oulgen

Differential Revision: D15045071

fbshipit-source-id: 6f0b038d7bf4b3ccbc78e8cec3a2377e75ac63aa