Track AddElem state in the interpreter
Summary:
This implements the Add*Elem functionality of AppendPeephole in the
interpreter. This avoids both the O(n^2) time and space complexity,
without requiring a special pass to do so, so we no longer need to
include AddElemC functions in the constant pass.

There's quite a bit of dead/redundant code after this, which I'll
clean up in follow up diffs.

Reviewed By: ricklavoie

Differential Revision: D15113035

fbshipit-source-id: c1bb7cfbd09a459280b6220b4de246b9c905ce56