Kill CollectionOpts::TrackConstantArrays
Summary:
This was used to prevent doing expensive array type tracking during
the constant pass analysis phase, since it would be redone during the
optimize phase.

Now that the analysis phase rewrites them anyway, we don't even need an
optimization phase for the constant pass anymore, so we can drop the flag.

Reviewed By: ricklavoie

Differential Revision: D15113882

fbshipit-source-id: eb446ecaef4796cfa9ca9ad496d30d3d441868a8