Use compact tagged pointer for the reified generics pointer on ActRec
Summary:
Whenever we call a function that takes reified generics, we need to check whether the generics provided are valid in arity and parity (erased/reified).
Prior to this diff, we do this by during the prologue calling a C++ function that iterates over the list of generics and compares  this to the data stored in ReifiedGenericsInfo. This is pretty inefficient. This diff improves the performance of this operation by using the higher order bits of the reified generics pointer (`m_reifiedGenerics`) to store a bitmap of whether each generic is reified or not. When we know the list of generics statically at the callsite (`SpillFrame`) we can compute bitmap statically and then at the prologue we need to test which is an assembly instruction rather than an expensive C++ call. When we do not know the generics statically, it is just as expensive.

Reviewed By: paulbiss

Differential Revision: D15056034

fbshipit-source-id: ce936eae910867b737b2c854a6876b9e019a3610