Don't rely on caches for producing decls on-demand
Summary: Currently, our APIs for getting decls compute it, then put it in a cache, and then expect the value to be there when retrieving it from the cache. This isn't guaranteed to work if the item was not saved to the cache (particularly in the serverless IDE case, where we want to limit the sizes of our caches). Instead, return the decl directly after we compute it, rather than reaching into the cache to get it.

Reviewed By: pittsw

Differential Revision: D14936032

fbshipit-source-id: 6179ecf40be2fa8ddbf408d3af38cbefed4e921a