extent allocator that manages a pre-allocated range with default fallback
Summary:
This intends to let an arena manage a preallocate piece of memory (probably backed by huge pages).

When the preallocated memory is used up, new mappings are added (and later purged) using the default allocation behavior, but the preallocated memory is never deallocated (until we later destroy the arena, if ever).

Reviewed By: interwq

Differential Revision: D15042165

fbshipit-source-id: d0e8234033ebc32e2413ce1ab1cd1406b075e096