use vm cold allocator for bytecode
Summary:
Make an allocator template for ReadOnlyArena, which uses the VMColdAllocator by default.

Killed the code that does mprotect(), because it isn't a good idea to make assumption about the underlying memory (it could be hugetlb pages). Doing mprotect each time we load new bytecode isn't good for efficiency anyway.

Also tried to hold the lock for a shorter time when allocating, in case multiple threads want to load bytecode concurrently.

Reviewed By: markw65

Differential Revision: D15033573

fbshipit-source-id: b31dc80c2cb194bea04d06496e68998d8d0ecd5c