Fix synthetic consts in legacy decl
Summary:
Every type constant has a matching synthetic class const that represents a TypeStructure. For abstract type constants, these class consts must be abstract. Previously, this was judged by the presence of a type on the type constant, but we can simplify this with the new ttc_abstract variant.

Also, when synthesizing a default of an abstract type constant, make its corresponding class const concrete.

Only applies to legacy decl

Differential Revision: D15043025

fbshipit-source-id: b7a94fd22632187c0857ea14ef9a370591b6bf95