fix startup hang
Summary: We are entering the debugger on onDefClass, which blocks the request in the thread that's loading the class. There's a global mutex in the VM protecting class loading, and since we block the thread waiting for commands, other threads trying to load classes block. This includes, sometimes, the dummy thread trying to run the startup document.

Reviewed By: ebluestein

Differential Revision: D15046960

fbshipit-source-id: 14a1d7cf3a1ccded9f71448554e120c99eec226b