Add ability to limit inlining by call-stack depth instead of stacked vasm cost
Summary: This diff adds a couple of options to control the inlining differently, so that the budget is adjusted based on the depth of the call instead of having a single budget for all the stacked calls. This is still experimental, so I'm leaving it off by default (so behavior should be unchanged).

Reviewed By: paulbiss

Differential Revision: D15151950

fbshipit-source-id: a85d95e267db480ee705eb9279cb79dc732123f0