New inference: simplify freshen_ty to just freshen all components of types
Summary:
When solving a type variable to its lower bound(s), we "freshen" the components of the types with type variables, and add explicitly any subtype consequences that follow. For example, if we have
```
vec<C> <: #1
```
we solve with `#1 := vec<#2>`, and add `C <: #2`.

The adding of subtype consequences is a premature optimization: it's much simpler to just freshen the components of the type, here obtaining `vec<#2>`, and then apply `sub_type` on `vec<C> <: vec<#2>` to obtain `C <: #2` by application of subtyping rules.

A forthcoming diff makes use of this freshening to construct a type in the "middle" of lower and upper bounds.

Reviewed By: CatherineGasnier

Differential Revision: D15147534

fbshipit-source-id: aef92f7286410b2d597fd8f8e76a4f843f4a0a7e