hh_force_solve pseudofunction
Summary:
This is used in a regression test in D15149670 where we want to make sure an environment is not growing out of control. At the end of the function, we write:
```
hh_force_solve()
hh_show($res)
```
and make sure the hh_show does not show a type that has grown too much.

Reviewed By: andrewjkennedy

Differential Revision: D15164576

fbshipit-source-id: a66254761de2ef7ece8fe29d489318b13212193c