Allow format strings to be appended as regular strings
Summary: The legacy type-checker treats format strings specially in that they are allowed to be used in contexts expecting strings (e.g., as operands of an append).  This diff ensures that the new inference algorithm allows that too.  The type of format strings, `HH\FormatString` is defined as an opaque newtype wrapper around string.  I don't know how important it is to keep it opaque.  An alternative way to achieve the same behavior as in this diff is to declare `HH\FormatString<T>` as a subtype of `string`.  (At the moment, format strings can be appended as regular strings but cannot be passed to functions expecting strings, which, IMO, is very confusing.)

Reviewed By: DavidSnider, andrewjkennedy

Differential Revision: D15147914

fbshipit-source-id: 3821cac8289ff0c6a0248d38e1954c28f4071d93