Allow frame eliding around memoization IR instructions
Summary:
The memoization IR instructions need the frame pointer in order to
read the keys (which are stored in locals). However they can do so
just as easily by using the stack pointer. Let the IR instructions
take either a frame pointer or stack pointer and calculate the offset
as appropriate. This allows us to elide the frame in memoization
wrappers.

Reviewed By: markw65

Differential Revision: D15169517

fbshipit-source-id: 64cccdc94e218b19ce152c5158e35f8367b3f368