Make the --config flag of hh_server actually work
Summary:
`hh_server` can be started with `--config` flag that is supposed to allow one to override the values of the config parameters set in `.hhconfig`.  For example, one should be able to set timeout as in
```
hh_server --check --config timeout=60 ~/www
```
Unfortunately, the implementation of this feature has a bug: we parse the overrides into a map and union it with the map of config parameters read from a config file.  The problem is that the union function that is used here is *left-biased*:

diffusion/FBS/browse/master/fbcode/hphp/hack/src/utils/collections/myMap.ml;c761edf5ab1902d157975ad63cc6145d28c3256d$25-30

i.e., if a key is present in both maps, the value from the left (first) map is chosen for the result, and so in

diffusion/FBS/browse/master/fbcode/hphp/hack/src/server/serverConfig.ml;c761edf5ab1902d157975ad63cc6145d28c3256d$225

`config_overrides` only has an effect for config parameters not set in `.hhconfig`.

Reviewed By: andrewjkennedy

Differential Revision: D15181422

fbshipit-source-id: ec4673847bbbc74537ed15fb5751a923eeb24059