Configure search provider for hh_single_type_check
Summary:
All programs that use the symbol index search provider must configure it on startup.

Prior to enabling the search provider to be used, I must modify hh_single_type_check so that it configures the symbol index provider on startup.

This code is necessary for other code that uses the search provider to work, so this must be lower in the diffstack.

Reviewed By: pittsw

Differential Revision: D15150531

fbshipit-source-id: e9c1da41fb2b9d81b781e4d4dbaf3407e5d00e46