Array provenance
Summary:
The idea here is to help track down where arrays come from when they're going
to be serialized, so we can log when we serialize, e.g. a v/darray that could
possibly change serialization formats as part of the HAM.

So--I keep a side table around with enough data to recover unit/line number and
set an extra bit in the ArrayData header if this data is present. The table is
cleared after every request and only contains request-allocated ArrayData's

In the appropriate minstrs (basically, any write minstr op), we sync VMregs and
store the frame pointer's Func* and PC--we can use these later to get the
unit/line number where the array was CoW'd.

Reviewed By: mxw

Differential Revision: D13982053

fbshipit-source-id: 099a63e1fd7c6153485e0d628ddc6e3587aabdf6