Don't punt on ConcatStr[34]
Summary:
When I moved the concat optimizations from peephole to interp, I made
them more aggressive, and we create a lot more ConcatN instructions.

I noticed that this caused a small instruction regression, and looking
at the jit for ConcatN it seems to be because we punt unless all the
inputs are known to be strings.

This adds the necessary conversions so we don't need to punt.

Reviewed By: oulgen

Differential Revision: D15188663

fbshipit-source-id: c001b97ac2c7e516f379f510cab04999324aed87