Add SQLite fallback for reverse naming table.
Summary: This change adds SQLite fallbacks for the position portion of the reverse naming table.  When attempting to read a value from the reverse naming table, we first check the shared memory tables, and if we cannot find the value there we read it from SQLite, save it back to shared memory (bringing the size of the tables from O(repo) to O(change+fanout)), and return the value.

Reviewed By: 2BitSalute

Differential Revision: D14674663

fbshipit-source-id: 3761dd3ede9a2670a2a16700d2876c9b50840131