Add SQLite fallback for the canon names table.
Summary:
This change adds a SQLite fallback to the second part of the reverse naming table, the canonical names table (canon table).  When we request a canonical name, we first check shared memory.  If it's not there, we look up the filename by using the canon name hash, then parse the file and grab the canonical name from that by doing a case insensitive comparison.  Once we have the canon name from the file, we save it back to the canon table.

We also add the canonical name to the shared memory canon table when we retrieve the position from SQLite.  This is safe to do, as the SQLite name key is case sensitive, which means that if we were able to retrieve a position we must have used the canonical name already.

Reviewed By: 2BitSalute

Differential Revision: D14674662

fbshipit-source-id: 85d16b80be7963674f75b6020e9b5e63e9c5d483