Abort in onBody if we have already sent a response.
Summary:
Presently HHVM will happily continue to buffer the body of a request even if we have already sent a response. This can happen e.g. if a misbehaving client just blindly sends body as part of a POST, not bothering to read a response midway.

Also adapt the handling of invalid content lengths to respond in a more useful and RFC-compliant way.

Additionally this adds enforcement of the max POST size into the onBody handling; as soon as a request body exceeds the post limit we respond with a 413.

Reviewed By: afrind

Differential Revision: D8211722

fbshipit-source-id: f6fbf4bb81af9bac21c56be4144b9e1a865c5399