Rename aast_* functions and clean up typing_get_locals.ml
Summary:
As part of the AST -> AAST migration, we also moved naming to operate over the AAST, mostly in anticipation of removing the legacy AST entirely.

However, as part of that migration, we left a bunch of functions prefaced with aast_* to distinguish them from the legacy versions. I've gone through and cleaned up the legacy implementations as much as I could and did a simple rename so that all of the code is visually more appealing.

Changes
- Change callsites of aast_lvar and aast_add_lvar
- Change uses of legacy_type_constraint
- Rename the implementation of fun_, class_, typedef and program
- Remove duplicates aast_lvar and aast_stmt from GetLocals module

Reviewed By: Wilfred

Differential Revision: D15105625

fbshipit-source-id: 41da96c3c81042bdb29e93efc14b271b7ee4c43c