Move mInstrState to CollectedInfo
Summary:
I was going to add another field to MInstrState, but realized the new
field didn't really belong in State, because it was only used within a
block.

Then I realized the same was true of MInstrState - its always empty at
block boundaries (and in fact state propagation ignores it, except
when the destination block is uninitialized), so we might as well keep
it in CollectedInfo, and avoid the space overhead of storing it per
block, and the cost of copying it when we propagate state.

We needed a few fixes for edge cases where we stop processing a block
in the middle of a minstr sequence.

Reviewed By: ricklavoie

Differential Revision: D15209737

fbshipit-source-id: f1720007ee5e635f1af68f35cdfe5e5e4b52aa48