Optimize minstr sequences in interp
Summary:
If a minstr sequence produces a literal value, we already have code in
dce to eliminate the sequence, but its complex, only catches the case
where the entire sequence produces a literal, and runs too late to get
the full benefits.

This moves the optimization to interp, and additionally catches cases
where any initial sequence produces a literal.

Reviewed By: ricklavoie

Differential Revision: D15209736

fbshipit-source-id: 6f533e1621efa83dd97e533558fa34e984fd58db