Delegate LazyTrivia lexer calls to Rust
Summary:
Spooky action at a distance time! Calls to `MinimalLexer` in `Full_fidelity_positioned_token.LazyTrivia` depend on global state set in Lexer during parsing. If we want to delegate those calls to Rust, we need a way to retrieve this global state and pass it to right function.

(Note: this ignores all push blocking failures!)

Reviewed By: pittsw

Differential Revision: D15094760

fbshipit-source-id: 3b30518ce3cd6434316f52c3e9675140457fcf1b