Add deletion to the reverse naming table fallback.
Summary: In order to initialize properly, we need to be able to selectively not write dirty files into the reverse naming table. This causes us problems when the reverse naming table is just a SQLite fallback, so the easiest way to support this is to remove all symbols from dirty files and then add deletion support to the reverse naming table when using the SQLite fallback.

Reviewed By: 2BitSalute

Differential Revision: D15053780

fbshipit-source-id: 8cb60fae8cad617149678db037f050ec8ac0af87