Flag errors whose primary position isn't the current file
Summary: For various reasons it's important that the "primary" position associated with an error (the first one in the raw formatting of the error) is located in the file that triggered the error. This diff introduces a hacky detection mechanism for when this fails to be the case. So far it has picked up one problem, with memoize type errors.

Reviewed By: manzyuk

Differential Revision: D15079343

fbshipit-source-id: 0265e1db14ba1c898990d5caeea044a69a7f4de2