Synthesize concrete type constants in new decl
Summary:
This diff synthesizes defaults into concrete type constants in the new decl paradigm. First, for each class's linearization we trace whether it is abstract and part of an unbroken chain of abstract type constants. This allows us to know when to synthesize type constants when lazily computing them.

```
abstract class A {
  abstract const type T as arraykey = int;
}
abstract class B extends A {
  abstract const type T as arraykey = string;
}
```
In both cases, we want to pass down the abstract type const as abstract.

```
abstract class A {
  abstract const type T as arraykey = int;
}
class B extends A {
  // const type T = int;
}
abstract class C extends B {
  abstract const type T as arraykey = string;
}
class D extends C {
  // const type T = string;
}
```
An important distinction to note here is that in the lazy decl, `D::T` will be a string rather than an `int` in the eager decl. In legacy decl, `B::T` is immediately synthesized into an `int` and all conflicts in child classes are dropped in favor of `B::T`.

Lazy decl `D` does not have direct access to `B::T`, so it looks up and synthesizes `C::T`. However, we are still safe as there is an error in C, during `Typing_extends`, in that it is trying to redeclare (synthetic) concrete type constant `B::T` as abstract.

Differential Revision: D15043189

fbshipit-source-id: 0597561e357fe21d996e007d6e207261f658d33e