Report all undefined types in one go
Summary:
When gen_type_scanners needs a type, but can't find a definition, it
throws an exception, reporting the name of the missing type.

As a result, if there are lots of missing types, you only find out
about the first one.

This changes it to report all the missing types in one go.

Reviewed By: binliu19, alexeyt

Differential Revision: D15242070

fbshipit-source-id: fbc9ef06a6ebb50fcdf7b83c3f7424693011c9c4