Try to dispatch method calls on known classes using their interfaces
Summary:
If we are trying to dispatch `$obj->name()` and we know that `$obj` is a
subclass of `C` that does not define `name()`, but implements `I` that defines
`name()`, we were taking slow path of `fpushObjMethodUnknown()`, which was
worse than what we would do if we knew just that `$obj` implements `I`.

Let's fix it by trying interface dispatch as well.

Also, do not check if the interface is unique, as we already know that the
object implements that specific interface, unique or not.

Reviewed By: ricklavoie

Differential Revision: D15192231

fbshipit-source-id: 6746be846bcda374f74e304ca270842b1f756390