Split ActRec spilling code paths into known, unknown and custom
Summary:
If a target Func* is known, use prepareToCallKnown(). If a target Func* is
unknown, but resolvable in JIT, use prepareToCallUnknown(). Otherwise, if we
look up function in C++ helper, which stores it directly to ActRec (potentially
with other ActRec fields), use prepareToCallCustom().

This refactoring makes it easier to move logic (such as caller checks, but
later the whole call dispatch) from FCall to the corresponding helpers.

Reviewed By: ricklavoie

Differential Revision: D15194506

fbshipit-source-id: acc0bcebe29eea9dd77051c0d2c5c3abdc0bc134