Indirection layer for autocomplete (Trie vs other providers)
Summary:
The goal of this diff is to allow a user to select legacy behavior or new behavior for autocomplete by specifying `--config symbolindex_search_provider=SqliteIndex` or something similar on the command line.

If that value is not provided, the user should not see changed behavior.

Since no other providers are implemented at the moment, the user should be able to see the behavior of "autocomplete symbol index disabled" if they set this value to anything other than `TrieIndex`.

Reviewed By: ljw1004

Differential Revision: D15177275

fbshipit-source-id: ae786cb5b6ad75e7cfb8a98bf0997d70017c44c5