Fold type_structure() when we statically know the result
Summary: This diff resolves the `type_structure(cls_name, const_name)` builtin statically at HHBBC time when we know the inputs statically. On a follow up diff, I'll also add the optimization to resolve it when the cls_name is `static::class` but the const_name is not overriden in children classes.

Reviewed By: markw65

Differential Revision: D15207877

fbshipit-source-id: 78d159b4968901c896401d791d74ebc2b68800e2