Fix a debug-types-section bug in debug-parser-dwarf
Summary:
With `-fdebug-types-section` enabled, a struct definition may simply
be a reference to a definition in a different type-unit. If we happen
to pick one of those as the canonical definition, it fails because it
looks like an definition (no declaration attribute), but is missing a
size attribute.

Just follow the reference if we find one.

Reviewed By: alexeyt

Differential Revision: D15251592

fbshipit-source-id: b1a2edfcd6625a750f0382f85ce410625960d5b6