New inference: solve for type variables with shallow-matching lower and upper bounds
Summary:
The function `bind_to_equal_bound` will currently solve a type variable if there is a type that appears *identically* in the its lower and upper bounds, for example, `vec<int> <: #1 <: vec<int>`. However, sometimes it is the case that we have bounds whose nested types differ but whose outer structure is the same, for example, `vec<int> <: #1 <: vec<arraykey>`. In this case it is useful to "partially" solve the type variable by freshening the nested types, e.g. `#1 := vec<#2>` and then apply subtyping e.g. `vec<int> <: vec<#2>` and `vec<#2> <: vec<arraykey>`.

The real utility of this change is through its use in `expand_type_and_narrow`, where we first apply `bind_to_equal_bound` to type variables in the type, in order to avoid unnecessary accumulation of new upper bounds, which can lead to very bad performance of type checing (TODO: find a small repro of this).

Reviewed By: CatherineGasnier, manzyuk

Differential Revision: D15238880

fbshipit-source-id: bac87f34ef90c15b49373067729684a9da6e9e79