Avoid `Lval on missing array element` logs from `IniSettingMap::set`
Summary:
The following command was resulting in stackless notices `Hack Array Compat: Lval on missing array element`:
```
/usr/local/hphpi/bin/hhvm -vEval.HackArrCompatNotices=1 -vEval.HackArrCompatCheckFalseyPromote=1 -m debug
```
I used `gdb` to trace this to two `lvalAt` calls via `IniSettingMap::set`.

I'm not sure if initializing the value at the missing key to null (`empty_array()`) is the best thing to but, it looks like the merging will just overwrite it.

Reviewed By: mofarrell

Differential Revision: D15202922

fbshipit-source-id: 354cd3e36f53a05aee33dc49208897413ac0dc73